Extract to separate delegate checker class
