+++ intellij-plugin-verifier/verifier-intellij/src/main/java/com/jetbrains/pluginverifier/usages/overrideOnly/AnActionUpdateMethodAllowedUsageFilter.kt
@@ -10,4 +10,4 @@ const val anActionClass: BinaryClassName = "com/intellij/openapi/actionSystem/An
 
 class AnActionUpdateMethodAllowedUsageFilter(
   private val delegate: ApiUsageFilter
-  = OverrideOnlyMethodAllowedUsageFilter(anActionUpdateMethodDescriptor, anActionClass)) : ApiUsageFilter by delegate
+  = OverrideOnlyMethodAllowedUsageFilter(anActionUpdateMethodDescriptor)) : ApiUsageFilter by delegate
